[+] refactor : card actions (clap + boo + share) component
